power actor tests part 3 fixes
