+++ actors/power/tests/power_actor_tests.rs
@@ -1,10 +1,10 @@
 use fil_actor_power::ext::init::{ExecParams, EXEC_METHOD};
 use fil_actor_power::ext::miner::MinerConstructorParams;
 use fil_actors_runtime::test_utils::{
-    expect_abort, MockRuntime, ACCOUNT_ACTOR_CODE_ID, CALLER_TYPES_SIGNABLE, MINER_ACTOR_CODE_ID,
+    expect_abort, ACCOUNT_ACTOR_CODE_ID, CALLER_TYPES_SIGNABLE, MINER_ACTOR_CODE_ID,
     SYSTEM_ACTOR_CODE_ID,
 };
-use fil_actors_runtime::{Multimap, INIT_ACTOR_ADDR};
+use fil_actors_runtime::INIT_ACTOR_ADDR;
 use fvm_ipld_encoding::{BytesDe, RawBytes};
 use fvm_shared::address::Address;
 use fvm_shared::clock::ChainEpoch;
@@ -14,8 +14,8 @@ use fvm_shared::sector::{RegisteredPoStProof, StoragePower};
 use num_traits::Zero;
 
 use fil_actor_power::{
-    epoch_key, Actor as PowerActor, CreateMinerParams, CronEvent, EnrollCronEventParams, Method,
-    State, UpdateClaimedPowerParams, CRON_QUEUE_AMT_BITWIDTH, CRON_QUEUE_HAMT_BITWIDTH,
+    Actor as PowerActor, CreateMinerParams, EnrollCronEventParams, Method, State,
+    UpdateClaimedPowerParams,
 };
 
 use crate::harness::*;
@@ -202,41 +202,6 @@ fn claimed_power_given_claim_does_not_exist_should_fail() {
     h.check_state();
 }
 
-fn get_enrolled_cron_ticks(rt: &MockRuntime, epoch: ChainEpoch) -> Vec<CronEvent> {
-    let state: State = rt.get_state();
-    let events_map = Multimap::from_root(
-        &rt.store,
-        &state.cron_event_queue,
-        CRON_QUEUE_HAMT_BITWIDTH,
-        CRON_QUEUE_AMT_BITWIDTH,
-    )
-    .expect("failed to load cron events");
-
-    let mut events: Vec<CronEvent> = Vec::new();
-    events_map
-        .for_each::<_, CronEvent>(&epoch_key(epoch), |_, v| {
-            events.push(v.to_owned());
-            Ok(())
-        })
-        .unwrap();
-
-    events
-}
-
-fn enroll_cron_event(
-    rt: &mut MockRuntime,
-    epoch: ChainEpoch,
-    miner_address: &Address,
-    payload: &RawBytes,
-) {
-    rt.set_caller(*MINER_ACTOR_CODE_ID, miner_address.to_owned());
-    rt.expect_validate_caller_type(vec![*MINER_ACTOR_CODE_ID]);
-    let params =
-        RawBytes::serialize(EnrollCronEventParams { event_epoch: epoch, payload: payload.clone() })
-            .unwrap();
-    rt.call::<PowerActor>(Method::EnrollCronEvent as u64, &params).unwrap();
-}
-
 #[test]
 fn enroll_cron_epoch_multiple_events() {
     let (h, mut rt) = setup();
@@ -244,21 +209,20 @@ fn enroll_cron_epoch_multiple_events() {
     let peer = "miner".as_bytes().to_vec();
     let multiaddrs = vec![BytesDe("multiaddr".as_bytes().to_vec())];
 
-    let miner1_address = Address::new_id(301);
     h.create_miner(
         &mut rt,
         &OWNER,
         &OWNER,
-        &miner1_address,
+        &MINER,
         &ACTOR,
         peer.clone(),
         multiaddrs.clone(),
         RegisteredPoStProof::StackedDRGWindow32GiBV1,
-        &TokenAmount::from(10),
+        &TokenAmount::zero(),
     )
     .unwrap();
 
-    let miner2_address = Address::new_id(302);
+    let miner2_address = Address::new_id(501);
     h.create_miner(
         &mut rt,
         &OWNER,
@@ -268,16 +232,16 @@ fn enroll_cron_epoch_multiple_events() {
         peer,
         multiaddrs,
         RegisteredPoStProof::StackedDRGWindow32GiBV1,
-        &TokenAmount::from(10),
+        &TokenAmount::zero(),
     )
     .unwrap();
 
     let mut enroll_and_check_cron_event = |epoch, miner_address, payload| {
-        let pre_existing_event_count = get_enrolled_cron_ticks(&rt, epoch).len();
+        let pre_existing_event_count = h.get_enrolled_cron_ticks(&rt, epoch).len();
 
-        enroll_cron_event(&mut rt, epoch, miner_address, payload);
+        h.enroll_cron_event(&mut rt, epoch, miner_address, payload);
 
-        let events = get_enrolled_cron_ticks(&rt, epoch);
+        let events = h.get_enrolled_cron_ticks(&rt, epoch);
         assert_eq!(events.len(), pre_existing_event_count + 1);
         assert_eq!(&events.last().unwrap().callback_payload, payload);
         assert_eq!(&events.last().unwrap().miner_addr, miner_address);
@@ -285,17 +249,16 @@ fn enroll_cron_epoch_multiple_events() {
 
     // enroll event with miner 1
     let payload = RawBytes::serialize(b"Cthulhu").unwrap();
-    enroll_and_check_cron_event(1, &miner1_address, &payload);
+    enroll_and_check_cron_event(1, &MINER, &payload);
 
     // enroll another event with the same miner
     let payload = RawBytes::serialize(b"Nyarlathotep").unwrap();
-    enroll_and_check_cron_event(1, &miner1_address, &payload);
+    enroll_and_check_cron_event(1, &MINER, &payload);
 
     // enroll another event with a different miner for a different epoch
     let payload = RawBytes::serialize(b"Azathoth").unwrap();
     enroll_and_check_cron_event(2, &miner2_address, &payload);
 
-    rt.verify();
     h.check_state();
 }
 
@@ -305,17 +268,16 @@ fn enroll_cron_epoch_before_current_epoch() {
 
     let peer = "miner".as_bytes().to_vec();
     let multiaddrs = vec![BytesDe("multiaddr".as_bytes().to_vec())];
-    let miner_address = Address::new_id(301);
     h.create_miner(
         &mut rt,
         &OWNER,
         &OWNER,
-        &miner_address,
+        &MINER,
         &ACTOR,
         peer,
         multiaddrs,
         RegisteredPoStProof::StackedDRGWindow32GiBV1,
-        &TokenAmount::from(10),
+        &TokenAmount::zero(),
     )
     .unwrap();
 
@@ -325,12 +287,12 @@ fn enroll_cron_epoch_before_current_epoch() {
     // enroll event with miner at epoch=2
     let miner_epoch = 2;
     let payload = RawBytes::serialize(b"Cthulhu").unwrap();
-    enroll_cron_event(&mut rt, miner_epoch, &miner_address, &payload);
+    h.enroll_cron_event(&mut rt, miner_epoch, &MINER, &payload);
 
-    let events = get_enrolled_cron_ticks(&rt, miner_epoch);
+    let events = h.get_enrolled_cron_ticks(&rt, miner_epoch);
     assert_eq!(events.len(), 1);
     assert_eq!(&events.last().unwrap().callback_payload, &payload);
-    assert_eq!(&events.last().unwrap().miner_addr, &miner_address);
+    assert_eq!(events.last().unwrap().miner_addr, *MINER);
 
     let state: State = rt.get_state();
     assert_eq!(state.first_cron_epoch, 0);
@@ -338,12 +300,12 @@ fn enroll_cron_epoch_before_current_epoch() {
     // enroll event with miner at epoch=1
     let miner_epoch = 1;
     let payload = RawBytes::serialize(b"Azathoth").unwrap();
-    enroll_cron_event(&mut rt, miner_epoch, &miner_address, &payload);
+    h.enroll_cron_event(&mut rt, miner_epoch, &MINER, &payload);
 
-    let events = get_enrolled_cron_ticks(&rt, miner_epoch);
+    let events = h.get_enrolled_cron_ticks(&rt, miner_epoch);
     assert_eq!(events.len(), 1);
     assert_eq!(&events.last().unwrap().callback_payload, &payload);
-    assert_eq!(&events.last().unwrap().miner_addr, &miner_address);
+    assert_eq!(events.last().unwrap().miner_addr, *MINER);
 
     let state: State = rt.get_state();
     assert_eq!(state.first_cron_epoch, 0);
